Fix integer overflow.
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene4547@1438455 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/util/packed/BlockPackedReader.java
@@ -45,7 +45,7 @@ public final class BlockPackedReader {
     blockShift = Integer.numberOfTrailingZeros(blockSize);
     blockMask = blockSize - 1;
     final int numBlocks = (int) (valueCount / blockSize) + (valueCount % blockSize == 0 ? 0 : 1);
-    if (numBlocks * blockSize < valueCount) {
+    if ((long) numBlocks * blockSize < valueCount) {
       throw new IllegalArgumentException("valueCount is too large for this block size");
     }
     long[] minValues = null;
